fix: 포인트 상한 설정 검증과 누적 합계 타입 보강

+++ src/main/kotlin/com/zerost/api/point/domain/PointHistoryRepository.kt
@@ -15,5 +15,5 @@ interface PointHistoryRepository : JpaRepository<PointHistory, Long> {
           and ph.amount > 0
         """,
     )
-    fun sumEarnedAmountByUserId(@Param("userId") userId: Long): Int
+    fun sumEarnedAmountByUserId(@Param("userId") userId: Long): Long
 }
